requirements.txt
cowsay==2.0.3
python-nmap==0.6.1
